bug(settings-menu): Menu item should work when present
Currently the settings menu item is not working on all views where
present.

This commit creates an activity with inflated menu resource and
all other activities that need to have this menu resource will just have
to inherit this activity.

[Fixes 157047945]
